chore: Clean up `dozer_sql::builder`
