+++ dozer-sql/src/lib.rs
@@ -2,7 +2,6 @@ mod aggregation;
 pub mod builder;
 pub mod errors;
 mod expression;
-mod pipeline_builder;
 mod planner;
 mod product;
 mod projection;
