--- dozer-sql/src/pipeline_builder/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod from_builder;
-pub(crate) mod join_builder;
